refactor(test): split tests in multiple files
